# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UMKM Website API",
        "description": "API Documentation for UMKM Website",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/businesses": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Get all businesses owned by the authenticated user",
                "operationId": "7123a3e6abeab779b8460888ff15e0fb",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "F&B"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter by city",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Malang"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of businesses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Kedai Es Teh Manis"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "F&B"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Malang"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Jl. Soekarno Hatta No. 12"
                                                    },
                                                    "latitude": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": -7.9666
                                                    },
                                                    "longitude": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 112.6326
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Business"
                ],
                "summary": "Create a new business",
                "operationId": "4533b15ef840def2c9ceb555a2bd66d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "city"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Kedai Es Teh Manis",
                                        "maxLength": 200
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "F&B",
                                        "maxLength": 100
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Malang",
                                        "maxLength": 100
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Soekarno Hatta No. 12"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -7.9666
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 112.6326
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Business created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Business created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/businesses/{id}": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Get a specific business by ID",
                "operationId": "fa4803adef18582b4d9e6728b83d6ca2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Business not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Business"
                ],
                "summary": "Update a specific business",
                "operationId": "d7697dd0261e3a4a072e3ba3cab52fbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Kedai Es Teh Manis Updated",
                                        "maxLength": 200
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "F&B",
                                        "maxLength": 100
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Surabaya",
                                        "maxLength": 100
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Basuki Rahmat No. 10"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -7.2575
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 112.7521
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Business updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Business not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Business"
                ],
                "summary": "Delete a specific business",
                "operationId": "5c158f0548ed580056aa06d26fd8a0a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Business deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Business not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}